mpls: Store and provide access to packet stats
It's useful to have statistics for mpls packets for a number of
reasons. RFC3813 details two packet stats that should be provided
(label lookup failures and fragmented packets) and also provides
interpretation of RFC2863 for other stats (in/out ucast, mcast and
bcast, in/out discards and errors and in unknown protos). Multicast
and fragment counters are provided for future implementation of
existing RFCs. Broadcast packet counters are printed, but not stored
to allow for possible future protocol extensions without the
user-space having to be changed to parse further stats.

All the introduced fields are 64-bit, even error ones, for simplicity
of code and to ensure no overflow even under network error
conditions. Per-CPU counters are used to avoid cache-line contention
on the commonly used fields. The other fields are similarly per-CPU
for simplicity.

The IPv6 proc code was used as an inspiration for the the proc code
here, both in terms of the implementation as well as the location of
the per-device stats proc files: /proc/net/dev_snmp_mpls/<dev>.