ipv6: storing and retrieval of per-nexthop encap
Parse the RTA_ENCAP attribute and store it in a newly added fields to
struct rt6_info, rt6i_encap and rt6i_encap_len. The encap is stored in
a separately allocated memory block.

Currently, the dst structure allocated to represent the destination of
the packet and used for retrieving the encap by the encap-type
interface has been grown through the addition of the rt6i_encap_len and
rt6i_encap fields. This isn't desirable and could be fixed by defining a
new destination type with operations copied from the normal case,
other than the addition of the get_encap operation.

Signed-off-by: Robert Shearman <rshearma@brocade.com>